drf_models/requirements.txt
anyio==4.0.0
AnyQt==0.2.0
asgiref==3.7.2
asttokens==2.4.0
attrs==23.1.0
backcall==0.2.0
baycomp==1.0.3
Bottleneck==1.3.7
catboost==1.2.2
cattrs==23.1.2
certifi==2023.7.22
chardet==5.2.0
charset-normalizer==3.3.0
colorama==0.4.6
comm==0.1.4
commonmark==0.9.1
contourpy==1.1.1
cycler==0.12.1
debugpy==1.8.0
decorator==5.1.1
dictdiffer==0.9.0
Django==4.2.5
django-cors-headers==4.3.0
django-restframework==0.0.1
djangorestframework==3.14.0
docutils==0.20.1
drf-yasg==1.21.7
et-xmlfile==1.1.0
executing==2.0.0
fonttools==4.43.1
graphviz==0.20.1
h11==0.14.0
httpcore==0.18.0
httpx==0.25.0
idna==3.4
importlib-metadata==6.8.0
inflection==0.5.1
ipykernel==6.25.2
ipython==8.16.1
ipython-genutils==0.2.0
jaraco.classes==3.3.0
jedi==0.19.1
joblib==1.3.2
jupyter_client==8.4.0
jupyter_core==5.4.0
keyring==24.2.0
keyrings.alt==5.0.0
kiwisolver==1.4.5
matplotlib==3.8.0
matplotlib-inline==0.1.6
more-itertools==10.1.0
nest-asyncio==1.5.8
networkx==3.1
numpy==1.26.1
openpyxl==3.1.2
openTSNE==1.0.0
orange-canvas-core==0.1.33
orange-widget-base==4.22.0
Orange3==3.36.1
packaging==23.2
pandas==2.1.1
parso==0.8.3
pickleshare==0.7.5
Pillow==10.1.0
platformdirs==3.11.0
plotly==5.17.0
prompt-toolkit==3.0.39
psutil==5.9.6
psycopg==3.1.12
psycopg-binary==3.1.12
psycopg-pool==3.1.8
pure-eval==0.2.2
Pygments==2.16.1
pyparsing==3.1.1
PyQt5==5.15.10
PyQt5-Qt5==5.15.2
PyQt5-sip==12.13.0
pyqtgraph==0.13.3
python-dateutil==2.8.2
python-louvain==0.16
pytz==2023.3.post1
pywin32==306
pywin32-ctypes==0.2.2
PyYAML==6.0.1
pyzmq==25.1.1
qasync==0.26.0
qtconsole==5.4.4
QtPy==2.4.0
requests==2.31.0
requests-cache==1.1.0
scikit-learn==1.1.3
scipy==1.11.3
serverfiles==0.3.1
six==1.16.0
sniffio==1.3.0
sqlparse==0.4.4
stack-data==0.6.3
tenacity==8.2.3
threadpoolctl==3.2.0
tornado==6.3.3
traitlets==5.11.2
typing_extensions==4.8.0
tzdata==2023.3
uritemplate==4.1.1
url-normalize==1.4.3
urllib3==2.0.7
wcwidth==0.2.8
xgboost==2.0.0
xlrd==2.0.1
XlsxWriter==3.1.8
zipp==3.17.0
